WIP: adding repository pattern
