+++ ui/package.json
@@ -34,8 +34,8 @@
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-svelte": "^2.35.1",
-    "jsdom": "^24.0.0",
     "globals": "^15.0.0",
+    "jsdom": "^24.0.0",
     "prettier": "^3.1.1",
     "prettier-plugin-svelte": "^3.1.2",
     "sass": "^1.77.2",
@@ -47,5 +47,8 @@
     "vite": "^5.0.3",
     "vitest": "^1.2.0"
   },
-  "type": "module"
+  "type": "module",
+  "dependencies": {
+    "kubernetes-types": "^1.30.0"
+  }
 }
