+++ ui/src/lib/repos/repository.ts
@@ -1,0 +1,6 @@
+import type{Pod} from 'kubernetes-types/core/v1';
+
+
+export interface PodRepository{
+    getPods(): Promise<Pod[]>;
+}
