+++ ui/src/routes/(platform)/+layout.ts
@@ -1,0 +1,7 @@
+import {FakePodRepository} from "../../../tests/fakerepo";
+
+const repo = new FakePodRepository();
+
+export async function load() {
+    return {repo}
+}
