+++ ui/src/routes/(platform)/cluster/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Cluster</h1>
